refactor: refactored transfer2 script
