+++ transfers/sensor_transfer.py
@@ -1,0 +1,29 @@
+# ===============================================================================
+# Copyright 2025 ross
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# ===============================================================================
+from datetime import datetime
+
+from db import Sensor
+
+
+# ============= EOF =============================================
+def init_sensor(session):
+    sensor = Sensor()
+    sensor.name = "Groundwater level manual measurement"
+    sensor.description = "manual gwl measurement. needs to be replaced with measurementmethod(?) e.g. steel tape, eprobe, etc."
+    sensor.unit = "ft"
+    sensor.datetime_installed = datetime.now()
+    session.add(sensor)
+    session.commit()
